+++ frontend/src/languages/fr.json
@@ -1139,5 +1139,9 @@
   "label.test.batch.cancel.finish": "Cette action est irréversible.",
   "label.test.batch.replace.start": "Les tests suivants",
   "label.test.batch.no.change.finish": "ne seront pas modifiés.",
-  "notification.slideover.button.unsubscribe": "Se désabonner"
+  "notification.slideover.button.unsubscribe": "Se désabonner",
+  "notification.slideover.button.unsubscribe.success": "Désabonnement réussi",
+  "notification.slideover.button.unsubscribe.fail": "Échec de l'annulation de l'abonnement",
+  "notification.slideover.button.subscribe.success": "Abonnement réussi",
+  "notification.slideover.button.subscribe.fail": "Échec de l'abonnement"
 }
