+++ src/main/java/org/openelisglobal/notifications/dao/NotificationSubscriptionDAO.java
@@ -12,4 +12,6 @@ public interface NotificationSubscriptionDAO {
 
     void saveOrUpdate(NotificationSubscriptions notificationSubscription);
 
+    void delete(NotificationSubscriptions ns);
+
 }
